+++ src/cpp_src/pb_methods_and_measures/GreedyOverCost.cpp
@@ -10,7 +10,7 @@
 #include <vector>
 
 std::vector<ProjectEmbedding> greedy_over_cost(const Election &election, const ProjectComparator &tie_breaking) {
-    int total_budget = election.budget();
+    long long total_budget = election.budget();
     auto projects = election.projects();
     std::vector<ProjectEmbedding> winners;
     std::ranges::sort(projects, [&tie_breaking](ProjectEmbedding a, ProjectEmbedding b) {
@@ -32,13 +32,13 @@ std::vector<ProjectEmbedding> greedy_over_cost(const Election &election, const P
     return winners;
 }
 
-std::optional<int> cost_reduction_for_greedy_over_cost(const Election &election, int p,
-                                                       const ProjectComparator &tie_breaking) {
-    int total_budget = election.budget();
+std::optional<long long> cost_reduction_for_greedy_over_cost(const Election &election, int p,
+                                                             const ProjectComparator &tie_breaking) {
+    long long total_budget = election.budget();
     auto projects = election.projects();
     auto pp = projects[p];
 
-    std::optional<int> max_price_to_be_chosen{};
+    std::optional<long long> max_price_to_be_chosen{};
 
     std::ranges::sort(projects, [&tie_breaking](ProjectEmbedding a, ProjectEmbedding b) {
         long long cross_term_a_approvals_b_cost = static_cast<long long>(a.approvers().size()) * b.cost(),
@@ -53,13 +53,13 @@ std::optional<int> cost_reduction_for_greedy_over_cost(const Election &election,
             if (project == pp) {
                 return pp.cost();
             } else {
-                int curr_max_price = 0;
+                long long curr_max_price = 0;
                 if (project.approvers().size() == 0) {
                     curr_max_price = project.cost();
                 } else {
-                    curr_max_price =
-                        std::min(static_cast<int>(project.cost() * pp.approvers().size() / project.approvers().size()),
-                                 total_budget); // todo: change if price doesn't have to be int
+                    curr_max_price = std::min(
+                        static_cast<long long>(project.cost() * pp.approvers().size() / project.approvers().size()),
+                        total_budget); // todo: change if price doesn't have to be long long
                 }
 
                 if (pp.approvers().size() * project.cost() == project.approvers().size() * curr_max_price &&
@@ -77,10 +77,10 @@ std::optional<int> cost_reduction_for_greedy_over_cost(const Election &election,
     return max_price_to_be_chosen;
 }
 
-std::optional<int> optimist_add_for_greedy_over_cost(const Election &election, int p,
-                                                     const ProjectComparator &tie_breaking) {
-    int total_budget = election.budget();
-    int num_voters = election.numVoters();
+std::optional<long long> optimist_add_for_greedy_over_cost(const Election &election, int p,
+                                                           const ProjectComparator &tie_breaking) {
+    long long total_budget = election.budget();
+    long long num_voters = election.numVoters();
     auto projects = election.projects();
     auto pp = projects[p];
     if (pp.cost() > total_budget)
@@ -101,7 +101,7 @@ std::optional<int> optimist_add_for_greedy_over_cost(const Election &election, i
                 return 0;
             }
             if (pp.cost() > total_budget - project.cost()) { // if (last moment to add pp)
-                int new_approvers_size =
+                long long new_approvers_size =
                     pbmath::ceil_div(static_cast<long long>(project.approvers().size()) * pp.cost(), project.cost());
                 std::vector<int> new_approvers(new_approvers_size);
                 std::iota(new_approvers.begin(), new_approvers.end(), 0);
@@ -123,14 +123,14 @@ std::optional<int> optimist_add_for_greedy_over_cost(const Election &election, i
     return {}; // LCOV_EXCL_LINE (every project should be feasible)
 }
 
-std::optional<int> pessimist_add_for_greedy_over_cost(const Election &election, int p,
-                                                      const ProjectComparator &tie_breaking) {
+std::optional<long long> pessimist_add_for_greedy_over_cost(const Election &election, int p,
+                                                            const ProjectComparator &tie_breaking) {
     return optimist_add_for_greedy_over_cost(election, p, tie_breaking);
 }
 
-std::optional<int> singleton_add_for_greedy_over_cost(const Election &election, int p,
-                                                      const ProjectComparator &tie_breaking) {
-    int total_budget = election.budget();
+std::optional<long long> singleton_add_for_greedy_over_cost(const Election &election, int p,
+                                                            const ProjectComparator &tie_breaking) {
+    long long total_budget = election.budget();
     auto projects = election.projects();
     auto pp = projects[p];
     if (pp.cost() > total_budget)
@@ -151,7 +151,7 @@ std::optional<int> singleton_add_for_greedy_over_cost(const Election &election, 
                 return 0;
             }
             if (pp.cost() > total_budget - project.cost()) { // if (last moment to add pp)
-                int new_approvers_size =
+                long long new_approvers_size =
                     pbmath::ceil_div(static_cast<long long>(project.approvers().size()) * pp.cost(), project.cost());
                 std::vector<int> new_approvers(new_approvers_size);
                 std::iota(new_approvers.begin(), new_approvers.end(), 0);
